Refactor Modal into it's own component
