+++ src/components/SavePointModal.tsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react';
+import Modal from './Modal';
 import './SavePointModal.css';
 
 interface SavePointModalProps {
@@ -37,43 +38,30 @@ export const SavePointModal: React.FC<SavePointModalProps> = ({
   if (!isOpen) return null;
 
   return (
-    <div className="modal-overlay" onClick={onClose}>
-      <div
-        className="save-point-modal"
-        onClick={e => e.stopPropagation()}
-        onKeyDown={handleKeyDown}
-      >
-        <div className="modal-header">
-          <h3>Save Point</h3>
-          <button className="modal-close" onClick={onClose}>
-            ×
-          </button>
+    <Modal isOpen={isOpen} onClose={onClose} title="Save Point" className="save-point-modal">
+      <form onSubmit={handleSubmit} className="save-point-form" onKeyDown={handleKeyDown}>
+        <div className="form-field">
+          <label htmlFor="point-label">Point Name:</label>
+          <input
+            id="point-label"
+            type="text"
+            value={label}
+            onChange={e => setLabel(e.target.value)}
+            placeholder="Enter a name for this point"
+            autoFocus
+            maxLength={50}
+          />
         </div>
 
-        <form onSubmit={handleSubmit} className="save-point-form">
-          <div className="form-field">
-            <label htmlFor="point-label">Point Name:</label>
-            <input
-              id="point-label"
-              type="text"
-              value={label}
-              onChange={e => setLabel(e.target.value)}
-              placeholder="Enter a name for this point"
-              autoFocus
-              maxLength={50}
-            />
-          </div>
-
-          <div className="modal-actions">
-            <button type="button" onClick={onClose} className="cancel-button">
-              Cancel
-            </button>
-            <button type="submit" disabled={!label.trim()} className="save-button">
-              Save Point
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
+        <div className="modal-actions">
+          <button type="button" onClick={onClose} className="cancel-button">
+            Cancel
+          </button>
+          <button type="submit" disabled={!label.trim()} className="save-button">
+            Save Point
+          </button>
+        </div>
+      </form>
+    </Modal>
   );
 };
