+++ src/components/VictoryModal.tsx
@@ -1,12 +1,12 @@
 import React, { useState, useMemo } from 'react';
-import { createPortal } from 'react-dom';
 import { useAppSelector, useAppDispatch } from '../store/hooks';
 import { togglePrivateKeyDisplayMode } from '../store/slices/uiSlice';
-import './VictoryModal.css';
 import { getPublicKeyFromPrivate } from '../utils/crypto.ts';
 import { generateShareMessage, shareMessage } from '../utils/gameUtils';
+import Modal from './Modal';
 import type { SavedPoint } from '../types/ecc.ts';
 import type { Challenge } from '../types/game.ts';
+import './VictoryModal.css';
 
 interface VictoryModalProps {
   isOpen: boolean;
@@ -94,112 +94,101 @@ export const VictoryModal: React.FC<VictoryModalProps> = ({
     }
   };
 
-  return createPortal(
-    <div className="victory-modal-overlay" onClick={onClose}>
-      <div className="victory-modal" onClick={e => e.stopPropagation()}>
-        <div className="victory-header">
-          <button className="victory-close-x" onClick={onClose} aria-label="Close">
-            ×
-          </button>
-          <div className="victory-icon">{gaveUp ? '🤷' : '🏆'}</div>
-          <h2 className="victory-title">{getVictoryTitle()}</h2>
-          <p className="victory-message">{getVictoryMessage()}</p>
+  return (
+    <Modal
+      isOpen={isOpen}
+      onClose={onClose}
+      title={`${gaveUp ? '🤷' : '🏆'} ${getVictoryTitle()}`}
+      className="victory-modal"
+    >
+      <p className="victory-message">{getVictoryMessage()}</p>
+      <div className="victory-stats">
+        <div className="stat-item">
+          <div className="stat-label">Wallet Address</div>
+          <div className="modal-value-container">
+            <span className="stat-value address-value" title={challenge.p2pkh_address}>
+              {challenge.p2pkh_address}
+            </span>
+          </div>
         </div>
-        <div className="victory-stats">
+
+        {!gaveUp && (
           <div className="stat-item">
-            <div className="stat-label">Wallet Address</div>
+            <div className="stat-label">Private Key</div>
             <div className="modal-value-container">
-              <span className="stat-value address-value" title={challenge.p2pkh_address}>
-                {challenge.p2pkh_address}
+              <span
+                className="stat-value address-value clickable"
+                onClick={() => dispatch(togglePrivateKeyDisplayMode())}
+                title={privateKeyHexMode ? 'Click to switch to decimal' : 'Click to switch to hex'}
+              >
+                {(() => {
+                  try {
+                    const keyValue = victoryPrivateKey.startsWith('0x')
+                      ? victoryPrivateKey.slice(2)
+                      : victoryPrivateKey;
+                    const keyBigInt = BigInt('0x' + keyValue);
+                    return privateKeyHexMode ? '0x' + keyBigInt.toString(16) : keyBigInt.toString();
+                  } catch {
+                    return victoryPrivateKey;
+                  }
+                })()}
               </span>
+              <button
+                className="copy-button"
+                onClick={() => {
+                  try {
+                    const keyValue = victoryPrivateKey.startsWith('0x')
+                      ? victoryPrivateKey.slice(2)
+                      : victoryPrivateKey;
+                    const keyBigInt = BigInt('0x' + keyValue);
+                    const textToCopy = privateKeyHexMode
+                      ? '0x' + keyBigInt.toString(16)
+                      : keyBigInt.toString();
+                    navigator.clipboard.writeText(textToCopy);
+                  } catch {
+                    navigator.clipboard.writeText(victoryPrivateKey);
+                  }
+                }}
+              >
+                Copy
+              </button>
             </div>
           </div>
+        )}
 
-          {!gaveUp && (
-            <div className="stat-item">
-              <div className="stat-label">Private Key</div>
-              <div className="modal-value-container">
-                <span
-                  className="stat-value address-value clickable"
-                  onClick={() => dispatch(togglePrivateKeyDisplayMode())}
-                  title={
-                    privateKeyHexMode ? 'Click to switch to decimal' : 'Click to switch to hex'
-                  }
-                >
-                  {(() => {
-                    try {
-                      const keyValue = victoryPrivateKey.startsWith('0x')
-                        ? victoryPrivateKey.slice(2)
-                        : victoryPrivateKey;
-                      const keyBigInt = BigInt('0x' + keyValue);
-                      return privateKeyHexMode
-                        ? '0x' + keyBigInt.toString(16)
-                        : keyBigInt.toString();
-                    } catch {
-                      return victoryPrivateKey;
-                    }
-                  })()}
-                </span>
-                <button
-                  className="copy-button"
-                  onClick={() => {
-                    try {
-                      const keyValue = victoryPrivateKey.startsWith('0x')
-                        ? victoryPrivateKey.slice(2)
-                        : victoryPrivateKey;
-                      const keyBigInt = BigInt('0x' + keyValue);
-                      const textToCopy = privateKeyHexMode
-                        ? '0x' + keyBigInt.toString(16)
-                        : keyBigInt.toString();
-                      navigator.clipboard.writeText(textToCopy);
-                    } catch {
-                      navigator.clipboard.writeText(victoryPrivateKey);
-                    }
-                  }}
-                >
-                  Copy
-                </button>
-              </div>
-            </div>
-          )}
+        <div className="stat-item">
+          <div className="stat-label">{gaveUp ? 'Point Operations tried' : 'Point Operations'}</div>
+          <div className="stat-value address-value">{operationCount}</div>
+        </div>
 
+        {signature && !gaveUp && (
           <div className="stat-item">
-            <div className="stat-label">
-              {gaveUp ? 'Point Operations tried' : 'Point Operations'}
+            <div className="stat-label">Bitcoin-cli Verification Command</div>
+            <div className="modal-value-container">
+              <span className="stat-value address-value">
+                {`bitcoin-cli verifymessage "${challenge.p2pkh_address}" "${signature}" "${victoryPublicKey}"`}
+              </span>
+              <button
+                className="copy-button"
+                onClick={() =>
+                  navigator.clipboard.writeText(
+                    `bitcoin-cli verifymessage "${challenge.p2pkh_address}" "${signature}" "${victoryPublicKey}"`
+                  )
+                }
+              >
+                Copy
+              </button>
             </div>
-            <div className="stat-value address-value">{operationCount}</div>
           </div>
-
-          {signature && !gaveUp && (
-            <div className="stat-item">
-              <div className="stat-label">Bitcoin-cli Verification Command</div>
-              <div className="modal-value-container">
-                <span className="stat-value address-value">
-                  {`bitcoin-cli verifymessage "${challenge.p2pkh_address}" "${signature}" "${victoryPublicKey}"`}
-                </span>
-                <button
-                  className="copy-button"
-                  onClick={() =>
-                    navigator.clipboard.writeText(
-                      `bitcoin-cli verifymessage "${challenge.p2pkh_address}" "${signature}" "${victoryPublicKey}"`
-                    )
-                  }
-                >
-                  Copy
-                </button>
-              </div>
-            </div>
-          )}
-        </div>
-
-        <div className="victory-actions">
-          <button onClick={handleShare} className="victory-share-button" autoFocus>
-            {shareStatus || 'Share Result'}
-          </button>
-        </div>
+        )}
       </div>
-    </div>,
-    document.body
+
+      <div className="victory-actions">
+        <button onClick={handleShare} className="victory-share-button" autoFocus>
+          {shareStatus || 'Share Result'}
+        </button>
+      </div>
+    </Modal>
   );
 };
